services/epr/EmployeePerformanceReport/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
numpy>=1.24.0
supabase>=1.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
mcp>=0.9.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pyjwt>=2.8.0
httpx>=0.25.0
python-multipart>=0.0.6
pydantic>=2.5.0

